pyproject.toml
[project]
name = "chative-core-llm-devmode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.10",
    "pymilvus>=2.6.2",
    "python-dotenv>=1.0.1",
]
